\name{regboost-package}
\alias{regboost-package}
\alias{regboost}
\docType{package}
\title{
\packageTitle{regboost}
}
\description{
\packageDescription{regboost}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{regboost}
\packageIndices{regboost}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{regboost}

Maintainer: \packageMaintainer{regboost}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
